feat: move EffectifDetail to shared folder
