+++ ui/app/_components/ruptures/MissionLocaleEffectifDisplay.tsx
