+++ ui/app/_components/ruptures/effectifs/ConfirmReset.tsx
