+++ ui/app/_components/ruptures/effectifs/EffectifInfo.module.css
