+++ ui/app/_components/ruptures/effectifs/EffectifInfo.tsx
