+++ ui/app/_components/ruptures/effectifs/EffectifInfoDetails.module.css
