+++ ui/app/_components/ruptures/effectifs/EffectifInfoDetails.tsx
