+++ ui/app/_components/ruptures/effectifs/RightColumnSkeleton.tsx
